Rename exoscale_database to exoscale_dbaas
dbaas: database management for mysql, pg

Wait after DB service is rebuilding

doc

typo

imports
