+++ docs/resources/dbaas.md
@@ -1,51 +1,16 @@
 ---
 # generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "exoscale_database Resource - terraform-provider-exoscale"
+page_title: "exoscale_dbaas Resource - terraform-provider-exoscale"
 subcategory: ""
 description: |-
-  ❗This resource is deprecated and renamed to exoscale_dbaas, do not use it to create new resources❗
   Manage Exoscale Database Services (DBaaS) https://community.exoscale.com/documentation/dbaas/.
 ---
 
-# exoscale_database (Resource)
+# exoscale_dbaas (Resource)
 
-❗This resource is deprecated and renamed to exoscale_dbaas, do not use it to create new resources❗
 Manage Exoscale [Database Services (DBaaS)](https://community.exoscale.com/documentation/dbaas/).
 
-## Example Usage
 
-```terraform
-resource "exoscale_database" "my_database" {
-  zone = "ch-gva-2"
-  name = "my-database"
-
-  type = "pg"
-  plan = "startup-4"
-
-  maintenance_dow  = "sunday"
-  maintenance_time = "23:00:00"
-
-  termination_protection = true
-
-  pg {
-    version = "13"
-
-    backup_schedule = "04:00"
-
-    ip_filter = [
-      "1.2.3.4/32",
-      "5.6.7.8/32",
-    ]
-
-    pg_settings = jsonencode({
-      timezone : "Europe/Zurich"
-    })
-  }
-}
-```
-
-Please refer to the [examples](https://github.com/exoscale/terraform-provider-exoscale/tree/master/examples/)
-directory for complete configuration examples.
 
 <!-- schema generated by tfplugindocs -->
 ## Schema
@@ -54,7 +19,7 @@ directory for complete configuration examples.
 
 - `name` (String) ❗ The name of the database service.
 - `plan` (String) The plan of the database service (use the [Exoscale CLI](https://github.com/exoscale/cli/) - `exo dbaas type show <TYPE> --plans` - for reference).
-- `type` (String) ❗ The type of the database service (`kafka`, `mysql`, `opensearch`, `pg`, `grafana`, `valkey`).
+- `type` (String) ❗ The type of the database service (`kafka`, `mysql`, `opensearch`, `pg`, `redis`, `grafana`).
 - `zone` (String) ❗ The Exoscale [Zone](https://www.exoscale.com/datacenters/) name.
 
 ### Optional
@@ -133,7 +98,7 @@ Optional:
 - `index_template` (Block, Optional) Template settings for all new indexes (see [below for nested schema](#nestedblock--opensearch--index_template))
 - `ip_filter` (Set of String) Allow incoming connections from this list of CIDR address block, e.g. `["10.20.0.0/16"]
 - `keep_index_refresh_interval` (Boolean) Aiven automation resets index.refresh_interval to default value for every index to be sure that indices are always visible to search. If it doesn't fit your case, you can disable this by setting up this flag to true.
-- `max_index_count` (Number, Deprecated) Maximum number of indexes to keep (Minimum value is `0`)
+- `max_index_count` (Number) Maximum number of indexes to keep (Minimum value is `0`)
 - `recovery_backup_name` (String) ❗ Name of a backup to recover from
 - `settings` (String) OpenSearch-specific settings, in json. e.g.`jsonencode({thread_pool_search_size: 64})`. Use `exo x get-dbaas-settings-opensearch` to get a list of available settings.
 - `version` (String) ❗ OpenSearch major version (`exo dbaas type show opensearch` for reference)
@@ -205,12 +170,4 @@ Optional:
 
 -> The symbol ❗ in an attribute indicates that modifying it, will force the creation of a new resource.
 
-## Import
 
-```shell
-# An existing database service may be imported by `<name>@<zone>`:
-
-terraform import \
-  exoscale_database.my_database \
-  my-database@ch-gva-2
-```
