+++ pkg/resources/database/resource_service_kafka.go
@@ -99,7 +99,7 @@ var ResourceKafkaSchema = schema.SingleNestedBlock{
 }
 
 // createKafka function handles Kafka specific part of database resource creation logic.
-func (r *Resource) createKafka(ctx context.Context, data *ResourceModel, diagnostics *diag.Diagnostics) {
+func (r *ServiceResource) createKafka(ctx context.Context, data *ServiceResourceModel, diagnostics *diag.Diagnostics) {
 	service := oapi.CreateDbaasServiceKafkaJSONRequestBody{
 		Plan:                  data.Plan.ValueString(),
 		TerminationProtection: data.TerminationProtection.ValueBoolPointer(),
@@ -210,7 +210,7 @@ func (r *Resource) createKafka(ctx context.Context, data *ResourceModel, diagnos
 
 // readKafka function handles Kafka specific part of database resource Read logic.
 // It is used in the dedicated Read action but also as a finishing step of Create, Update and Import.
-func (r *Resource) readKafka(ctx context.Context, data *ResourceModel, diagnostics *diag.Diagnostics) {
+func (r *ServiceResource) readKafka(ctx context.Context, data *ServiceResourceModel, diagnostics *diag.Diagnostics) {
 	caCert, err := r.client.GetDatabaseCACertificate(ctx, data.Zone.ValueString())
 	if err != nil {
 		diagnostics.AddError("Client Error", fmt.Sprintf("Unable to get CA Certificate: %s", err))
@@ -323,7 +323,7 @@ func (r *Resource) readKafka(ctx context.Context, data *ResourceModel, diagnosti
 }
 
 // updateKafka function handles Kafka specific part of database resource Update logic.
-func (r *Resource) updateKafka(ctx context.Context, stateData *ResourceModel, planData *ResourceModel, diagnostics *diag.Diagnostics) {
+func (r *ServiceResource) updateKafka(ctx context.Context, stateData *ServiceResourceModel, planData *ServiceResourceModel, diagnostics *diag.Diagnostics) {
 	var updated bool
 
 	service := oapi.UpdateDbaasServiceKafkaJSONRequestBody{}
