+++ pkg/resources/database/resource_service_pg.go
@@ -80,7 +80,7 @@ var ResourcePgSchema = schema.SingleNestedBlock{
 }
 
 // createPg function handles PostgreSQL specific part of database resource creation logic.
-func (r *Resource) createPg(ctx context.Context, data *ResourceModel, diagnostics *diag.Diagnostics) {
+func (r *ServiceResource) createPg(ctx context.Context, data *ServiceResourceModel, diagnostics *diag.Diagnostics) {
 	service := oapi.CreateDbaasServicePgJSONRequestBody{
 		Plan:                  data.Plan.ValueString(),
 		TerminationProtection: data.TerminationProtection.ValueBoolPointer(),
@@ -302,7 +302,7 @@ func (r *Resource) createPg(ctx context.Context, data *ResourceModel, diagnostic
 
 // readPg function handles PostgreSQL specific part of database resource Read logic.
 // It is used in the dedicated Read action but also as a finishing step of Create, Update and Import.
-func (r *Resource) readPg(ctx context.Context, data *ResourceModel, diagnostics *diag.Diagnostics) {
+func (r *ServiceResource) readPg(ctx context.Context, data *ServiceResourceModel, diagnostics *diag.Diagnostics) {
 	caCert, err := r.client.GetDatabaseCACertificate(ctx, data.Zone.ValueString())
 	if err != nil {
 		diagnostics.AddError("Client Error", fmt.Sprintf("Unable to get CA Certificate: %s", err))
@@ -461,7 +461,7 @@ func (r *Resource) readPg(ctx context.Context, data *ResourceModel, diagnostics 
 }
 
 // updatePg function handles PostgreSQL specific part of database resource Update logic.
-func (r *Resource) updatePg(ctx context.Context, stateData *ResourceModel, planData *ResourceModel, diagnostics *diag.Diagnostics) {
+func (r *ServiceResource) updatePg(ctx context.Context, stateData *ServiceResourceModel, planData *ServiceResourceModel, diagnostics *diag.Diagnostics) {
 	var updated bool
 
 	service := oapi.UpdateDbaasServicePgJSONRequestBody{}
