+++ pkg/resources/database/resource_user_pg.go
@@ -4,8 +4,9 @@ import (
 	"context"
 	"fmt"
 	"strings"
+	"time"
 
-	exoscale "github.com/exoscale/egoscale/v3"
+	v3 "github.com/exoscale/egoscale/v3"
 	providerConfig "github.com/exoscale/terraform-provider-exoscale/pkg/provider/config"
 	"github.com/hashicorp/terraform-plugin-framework-timeouts/resource/timeouts"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
@@ -67,23 +68,23 @@ func (r *PGUserResource) Schema(ctx context.Context, req resource.SchemaRequest,
 
 func (r *PGUserResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
 	var data PGUserResourceModel
-	UserRead(ctx, req, resp, &data, r.client)
+	ReadResource(ctx, req, resp, &data, r.client)
 }
 
 func (r *PGUserResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 
 	var data PGUserResourceModel
-	UserCreate(ctx, req, resp, &data, r.client)
+	CreateResource(ctx, req, resp, &data, r.client)
 }
 
 func (r *PGUserResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
 	var stateData, planData PGUserResourceModel
-	UserUpdate(ctx, req, resp, &stateData, &planData, r.client)
+	UpdateResource(ctx, req, resp, &stateData, &planData, r.client)
 }
 
 func (r *PGUserResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
 	var data PGUserResourceModel
-	UserDelete(ctx, req, resp, &data, r.client)
+	DeleteResource(ctx, req, resp, &data, r.client)
 }
 
 func (r *PGUserResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
@@ -130,16 +131,15 @@ func (r *PGUserResource) ImportState(ctx context.Context, req resource.ImportSta
 	data.Username = types.StringValue(username)
 	data.Service = types.StringValue(serviceName)
 	data.Zone = types.StringValue(zone)
-	data.Zone = types.StringValue(zone)
 
-	UserReadForImport(ctx, req, resp, &data, r.client)
+	ReadResourceForImport(ctx, req, resp, &data, r.client)
 
 }
 
-func (data *PGUserResourceModel) CreateResource(ctx context.Context, client *exoscale.Client, diagnostics *diag.Diagnostics) {
+func (data *PGUserResourceModel) CreateResource(ctx context.Context, client *v3.Client, diagnostics *diag.Diagnostics) {
 
-	createRequest := exoscale.CreateDBAASPostgresUserRequest{
-		Username: exoscale.DBAASUserUsername(data.Username.ValueString()),
+	createRequest := v3.CreateDBAASPostgresUserRequest{
+		Username: v3.DBAASUserUsername(data.Username.ValueString()),
 	}
 
 	if !data.AllowReplication.IsNull() {
@@ -155,7 +155,7 @@ func (data *PGUserResourceModel) CreateResource(ctx context.Context, client *exo
 		return
 	}
 
-	_, err = client.Wait(ctx, op, exoscale.OperationStateSuccess)
+	_, err = client.Wait(ctx, op, v3.OperationStateSuccess)
 	if err != nil {
 		diagnostics.AddError(
 			"Client Error",
@@ -191,7 +191,7 @@ func (data *PGUserResourceModel) CreateResource(ctx context.Context, client *exo
 	diagnostics.AddError("Client Error", "Unable to find newly created user for the service")
 }
 
-func (data *PGUserResourceModel) Delete(ctx context.Context, client *exoscale.Client, diagnostics *diag.Diagnostics) {
+func (data *PGUserResourceModel) DeleteResource(ctx context.Context, client *v3.Client, diagnostics *diag.Diagnostics) {
 
 	op, err := client.DeleteDBAASPostgresUser(ctx, data.Service.ValueString(), data.Username.ValueString())
 	if err != nil {
@@ -202,7 +202,7 @@ func (data *PGUserResourceModel) Delete(ctx context.Context, client *exoscale.Cl
 		return
 	}
 
-	_, err = client.Wait(ctx, op, exoscale.OperationStateSuccess)
+	_, err = client.Wait(ctx, op, v3.OperationStateSuccess)
 	if err != nil {
 		diagnostics.AddError(
 			"Client Error",
@@ -213,7 +213,8 @@ func (data *PGUserResourceModel) Delete(ctx context.Context, client *exoscale.Cl
 
 }
 
-func (data *PGUserResourceModel) ReadResource(ctx context.Context, client *exoscale.Client, diagnostics *diag.Diagnostics) {
+func (data *PGUserResourceModel) ReadResource(ctx context.Context, client *v3.Client, diagnostics *diag.Diagnostics) {
+
 	svc, err := client.GetDBAASServicePG(ctx, data.Service.ValueString())
 	if err != nil {
 		diagnostics.AddError("Client Error", fmt.Sprintf("Unable to read service pg user, got error: %s", err))
@@ -241,15 +242,18 @@ func (data *PGUserResourceModel) ReadResource(ctx context.Context, client *exosc
 	diagnostics.AddError("Client Error", "Unable to read user for the service")
 }
 
-func (data *PGUserResourceModel) UpdateResource(ctx context.Context, client *exoscale.Client, diagnostics *diag.Diagnostics) {
+func (data *PGUserResourceModel) UpdateResource(ctx context.Context, client *v3.Client, diagnostics *diag.Diagnostics) {
 	// Nothing to do here as all fields of this resource are immutable; replaces will be required
 	// automatically
 }
 
-func (data *PGUserResourceModel) WaitForService(ctx context.Context, client *exoscale.Client, diagnostics *diag.Diagnostics) {
-	_, err := waitForDBAASServiceReadyForUsers(ctx, client.GetDBAASServicePG, data.Service.ValueString(), func(t *exoscale.DBAASServicePG) bool { return len(t.Users) > 0 })
+func (data *PGUserResourceModel) WaitForService(ctx context.Context, client *v3.Client, diagnostics *diag.Diagnostics) {
+	_, err := waitForDBAASServiceReadyForFn(ctx, client.GetDBAASServicePG, data.Service.ValueString(), func(t *v3.DBAASServicePG) bool { return t.State == v3.EnumServiceStateRunning })
+	// DbaaS API is unstable when a service goes from rebuilding from running,
+	// this wait time helps avoid that
+	time.Sleep(time.Second * 10)
 	if err != nil {
-		diagnostics.AddError("Client Error", fmt.Sprintf("Unable to read Database service Opensearch %s", err.Error()))
+		diagnostics.AddError("Client Error", fmt.Sprintf("Unable to read Database service PG %s", err.Error()))
 	}
 }
