+++ pkg/resources/database/testdata/resource_kafka.tmpl
@@ -1,4 +1,4 @@
-resource "exoscale_database" {{ .ResourceName }} {
+resource "exoscale_dbaas" {{ .ResourceName }} {
   name = "{{ .Name }}"
   type = "kafka"
   plan = "{{ .Plan }}"
