+++ pkg/resources/database/testdata/resource_mysql.tmpl
@@ -1,4 +1,4 @@
-resource "exoscale_database" {{ .ResourceName }} {
+resource "exoscale_dbaas" {{ .ResourceName }} {
   name = "{{ .Name }}"
   type = "mysql"
   plan = "{{ .Plan }}"
